backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.122.0
uvicorn==0.38.0
pydantic-settings>=2.1.0
pydantic[email]>=2.0.0  # For EmailStr validation
python-multipart>=0.0.6

# AI & Vector DB dependencies
google-generativeai>=0.8.0
qdrant-client>=1.16.0

# Database dependencies (required for production)
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0  # Use psycopg 3.x with binary (better Python 3.13 support)
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4